refactor(config): move inline tests to module
Move the current codewhale-config inline test module into crates/config/src/tests.rs and leave crates/config/src/lib.rs with a small #[cfg(test)] module declaration. No production logic changed.

Harvested from PR #3345 by @cyq1017; the PR branch was stale against this integration branch, so this commit re-applies the same extraction to the current test body instead of cherry-picking outdated content.

Verification:\n- cargo fmt --all -- --check\n- git diff --check\n- cargo test -p codewhale-config --locked
